--- webapp/loggedin/modules/data/dataVis/dataVisView.js
@@ -1,12 +1,0 @@
-import './dataVisView.scss'
-import React from 'react'
-
-import DataQueryView from './dataQuery'
-
-const DataVisView = () => (
-  <div className="data-vis">
-    <DataQueryView />
-  </div>
-)
-
-export default DataVisView
