--- webapp/loggedin/modules/data/dataVis/dataVisView.scss
@@ -1,3 +1,0 @@
-.data-vis {
-  display: grid;
-}
