+++ webapp/loggedin/modules/data/reducer.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux'
 
-import dataVis from './dataVis/reducer'
+import dataVis from '../../../views/App/views/Data/Explorer/reducer'
 
 export default combineReducers({
   dataVis,
