+++ webapp/views/App/views/Data/Explorer/Explorer.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import DataQueryView from './dataQuery'
+
+const Explorer = () => {
+  return <DataQueryView />
+}
+
+export default Explorer
