+++ webapp/views/App/views/Data/Explorer/actions.js
