+++ webapp/views/App/views/Data/Explorer/dataQuery/actions/fetch.js
@@ -2,7 +2,7 @@ import axios from 'axios'
 
 import * as DataSort from '@common/surveyRdb/dataSort'
 import { SurveyState } from '@webapp/store/survey'
-import * as DataQueryState from '@webapp/loggedin/modules/data/dataVis/dataQuery/state'
+import * as DataQueryState from '@webapp/views/App/views/Data/Explorer/dataQuery/state'
 
 export const dataQueryTableInit = 'dataQuery/table/init'
