+++ webapp/views/App/views/Data/Explorer/dataQuery/components/dataQueryView.scss
