+++ webapp/views/App/views/Data/Explorer/dataQuery/components/sort/sortEditor.scss
@@ -1,4 +1,4 @@
-@import "~@webapp/style/vars";
+@import 'webapp/style/vars';
 
 .sort-editor-popup {
   .sort-editor__criteria {
@@ -6,7 +6,7 @@
 
     .sort-row {
       display: flex;
-      margin-bottom: .5rem;
+      margin-bottom: 0.5rem;
 
       .sort-row__label {
         width: 6rem;
@@ -18,7 +18,7 @@
         padding-left: 1rem;
 
         .btns__last {
-          margin-left: .6rem;
+          margin-left: 0.6rem;
         }
       }
     }
@@ -29,7 +29,7 @@
     align-self: end;
 
     > button {
-      margin: 0px .3rem;
+      margin: 0px 0.3rem;
       font-weight: 600;
     }
   }
