+++ webapp/views/App/views/Data/Explorer/dataQuery/components/table.js
