+++ webapp/views/App/views/Data/Explorer/dataQuery/components/tableContent/tableColumnEdit.js
