+++ webapp/views/App/views/Data/Explorer/dataQuery/components/tableHeader/tableHeader.scss
