+++ webapp/views/App/views/Data/Explorer/dataQuery/state/defaults.js
