+++ webapp/views/App/views/Data/Explorer/dataQuery/state/keys.js
