+++ webapp/views/App/views/Data/Explorer/dataQuery/state/read.js
@@ -1,4 +1,4 @@
 import * as R from 'ramda'
-import * as DataVisState from '@webapp/loggedin/modules/data/dataVis/state'
+import * as DataVisState from '@webapp/views/App/views/Data/Explorer/state'
 
 export const getState = R.pipe(DataVisState.getState, R.prop('query'))
