+++ webapp/views/App/views/Data/Explorer/dataQuery/state/table.js
