+++ webapp/views/App/views/Data/Explorer/reducer.js
